AgenticAI-ResearchAssistant/requirements1.txt
langchain==0.2.14
langchain-groq==0.1.6 
tavily-python==0.5.0
python-dotenv==1.0.1
